Add tests for streaming (3-arg) functions
