+++ src/test/php/com/amazon/aws/lambda/unittest/TestStream.class.php
@@ -1,0 +1,15 @@
+<?php namespace com\amazon\aws\lambda\unittest;
+
+use com\amazon\aws\lambda\Stream;
+
+class TestStream implements Stream {
+  public $written= '';
+
+  public function transmit($source, $mime= null) { /* NOOP */ }
+
+  public function use($mime) { /** NOOP */ }
+
+  public function write($bytes) { $this->written.= $bytes; }
+
+  public function end() { /** NOOP */ }
+}
